pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Jonatas E S Silva <jonatassilvaemanuel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
alembic = "^1.12.0"
python-decouple = "^3.8"
pydantic-settings = "^2.0.3"
pydantic = "^2.4.2"
sqlalchemy = "^2.0.22"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
